+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.0-beta.1
+- Added input binding
+- Added prevent default support
+- Added support for rendering actions without state using the `nostate` attribute
+
 ## 0.1.0-beta
 - Added Support for Event Handling
 - Added basic state management support